Move classes around to new packages.
