+++ src/main/java/com/mikeleitz/sidekick/bash/domain/BashProcessing.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mikeleitz.sidekick.bash;
+package com.mikeleitz.sidekick.bash.domain;
 
 /**
  * @author leitz@mikeleitz.com
